+++ BudgetApp/assets/adaptive-icon.png
