+++ BudgetApp/components/ErrorIndicator.js
