+++ BudgetApp/styles/globalStyles.js
