--- ฺBudgetApp/App.js
@@ -1,68 +1,0 @@
-import 'react-native-gesture-handler';
-import React, { useCallback, useEffect, useState } from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { GestureHandlerRootView } from 'react-native-gesture-handler';
-import { StyleSheet, View } from 'react-native';
-import { SafeAreaProvider } from 'react-native-safe-area-context';
-import * as SplashScreen from 'expo-splash-screen';
-import * as Font from 'expo-font';
-import AppNavigator from './navigation/AppNavigator';
-import { PocketProvider } from './context/PocketContext';
-import { UserProvider } from './context/UserContext';
-
-// Keep the splash screen visible while we fetch resources
-SplashScreen.preventAutoHideAsync();
-
-export default function App() {
-  const [appIsReady, setAppIsReady] = useState(false);
-
-  useEffect(() => {
-    async function prepare() {
-      try {
-        // Pre-load fonts
-        await Font.loadAsync({
-          // Load Ionicons font
-          ...require('@expo/vector-icons/build/vendor/react-native-vector-icons/Fonts/Ionicons.ttf'),
-        });
-      } catch (e) {
-        console.warn('Error loading fonts:', e);
-      } finally {
-        setAppIsReady(true);
-      }
-    }
-
-    prepare();
-  }, []);
-
-  const onLayoutRootView = useCallback(async () => {
-    if (appIsReady) {
-      await SplashScreen.hideAsync();
-    }
-  }, [appIsReady]);
-
-  if (!appIsReady) {
-    return null;
-  }
-
-  return (
-    <GestureHandlerRootView style={styles.container}>
-      <View style={styles.container} onLayout={onLayoutRootView}>
-        <SafeAreaProvider>
-          <UserProvider>
-            <PocketProvider>
-              <NavigationContainer>
-                <AppNavigator />
-              </NavigationContainer>
-            </PocketProvider>
-          </UserProvider>
-        </SafeAreaProvider>
-      </View>
-    </GestureHandlerRootView>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
