--- ฺBudgetApp/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "budgetapp",
-  "license": "0BSD",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web"
-  },
-  "dependencies": {
-    "@expo/metro-runtime": "~4.0.1",
-    "@expo/vector-icons": "^14.0.4",
-    "@react-native-async-storage/async-storage": "^1.23.1",
-    "@react-native-community/datetimepicker": "^8.3.0",
-    "@react-native-firebase/app": "^21.11.0",
-    "@react-native-firebase/auth": "^21.11.0",
-    "@react-native-firebase/firestore": "^21.11.0",
-    "@react-native-masked-view/masked-view": "^0.3.2",
-    "@react-navigation/bottom-tabs": "^7.2.1",
-    "@react-navigation/native": "^7.0.15",
-    "@react-navigation/native-stack": "^7.2.1",
-    "@react-navigation/stack": "^7.1.2",
-    "@shopify/react-native-skia": "1.5.0",
-    "axios": "^1.8.1",
-    "expo": "~52.0.37",
-    "expo-font": "~13.0.4",
-    "expo-splash-screen": "~0.29.3",
-    "expo-status-bar": "~2.0.1",
-    "firebase": "^11.4.0",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-native": "0.76.7",
-    "react-native-gesture-handler": "^2.24.0",
-    "react-native-linear-gradient": "^2.8.3",
-    "react-native-reanimated": "^3.17.1",
-    "react-native-safe-area-context": "^4.14.1",
-    "react-native-screens": "~4.4.0",
-    "react-native-svg": "^15.11.2",
-    "react-native-vector-icons": "^10.2.0",
-    "react-native-web": "~0.19.13",
-    "victory-native": "^41.16.1"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.20.0",
-    "@react-native-community/cli": "latest"
-  },
-  "private": true
-}
-
