--- ฺBudgetApp/screens/AddTransaction.js
@@ -1,413 +1,0 @@
-import React, { useState } from 'react';
-import {
-  View,
-  Text,
-  StyleSheet,
-  TextInput,
-  TouchableOpacity,
-  KeyboardAvoidingView,
-  Platform,
-  Alert,
-  ActivityIndicator,
-  ScrollView,
-  Switch,
-  Modal,
-} from 'react-native';
-import { SafeAreaView } from 'react-native-safe-area-context';
-import { usePockets } from '../context/PocketContext';
-import { Ionicons } from '@expo/vector-icons';
-import Icon from 'react-native-vector-icons/Ionicons';
-import { useNavigation } from '@react-navigation/native';
-
-export default function AddTransaction({ route, navigation }) {
-  const { pocketId, category, onTransactionAdded } = route.params;
-  const { addTransaction, refreshPockets } = usePockets();
-  const [amount, setAmount] = useState('');
-  const [description, setDescription] = useState('');
-  const [showDateModal, setShowDateModal] = useState(false);
-  const [selectedDate, setSelectedDate] = useState(new Date());
-  const [isRecurring, setIsRecurring] = useState(false);
-  const [recurringPeriod, setRecurringPeriod] = useState('monthly');
-  const [paymentMethod, setPaymentMethod] = useState('cash');
-  const [tags, setTags] = useState([]);
-  const [type, setType] = useState('expense');
-  const [loading, setLoading] = useState(false);
-
-  const handleAddTransaction = async () => {
-    if (!amount || !description) {
-      Alert.alert('Error', 'Please fill in all fields');
-      return;
-    }
-
-    try {
-      setLoading(true);
-      await addTransaction(pocketId, {
-        type,
-        amount: parseFloat(amount),
-        description,
-        date: new Date().toISOString(),
-      });
-
-      await refreshPockets();
-
-      if (onTransactionAdded) {
-        onTransactionAdded();
-      }
-
-      navigation.goBack();
-    } catch (error) {
-      Alert.alert('Error', error.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const renderDatePicker = () => {
-    const today = new Date();
-    const dates = [];
-    for (let i = 0; i < 7; i++) {
-      const date = new Date();
-      date.setDate(today.getDate() - i);
-      dates.push(date);
-    }
-
-    return (
-      <Modal
-        visible={showDateModal}
-        transparent={true}
-        animationType="slide"
-        onRequestClose={() => setShowDateModal(false)}
-      >
-        <View style={styles.modalOverlay}>
-          <View style={styles.modalContent}>
-            <Text style={styles.modalTitle}>Select Date</Text>
-            <ScrollView>
-              {dates.map((date) => (
-                <TouchableOpacity
-                  key={date.toISOString()}
-                  style={[
-                    styles.dateOption,
-                    date.toDateString() === selectedDate.toDateString() && styles.selectedDate
-                  ]}
-                  onPress={() => {
-                    setSelectedDate(date);
-                    setShowDateModal(false);
-                  }}
-                >
-                  <Text style={[
-                    styles.dateText,
-                    date.toDateString() === selectedDate.toDateString() && styles.selectedDateText
-                  ]}>
-                    {date.toLocaleDateString('en-US', {
-                      weekday: 'short',
-                      month: 'short',
-                      day: 'numeric'
-                    })}
-                  </Text>
-                </TouchableOpacity>
-              ))}
-            </ScrollView>
-            <TouchableOpacity
-              style={styles.closeButton}
-              onPress={() => setShowDateModal(false)}
-            >
-              <Text style={styles.closeButtonText}>Close</Text>
-            </TouchableOpacity>
-          </View>
-        </View>
-      </Modal>
-    );
-  };
-
-  return (
-    <SafeAreaView style={styles.container}>
-      <View style={styles.header}>
-        <TouchableOpacity
-          onPress={() => navigation.goBack()}
-          style={styles.backButton}
-        >
-          <Icon name="arrow-back" size={24} color="#000" />
-        </TouchableOpacity>
-        <Text style={styles.headerTitle}>Add Transaction</Text>
-        <View style={styles.rightPlaceholder} />
-      </View>
-
-      <ScrollView style={styles.content}>
-        <View style={styles.typeSelector}>
-          <TouchableOpacity
-            style={[styles.typeButton, type === 'expense' && styles.activeTypeButton]}
-            onPress={() => setType('expense')}
-          >
-            <Text style={[styles.typeButtonText, type === 'expense' && styles.activeTypeText]}>
-              Expense
-            </Text>
-          </TouchableOpacity>
-          {category !== 'bills' && (
-            <TouchableOpacity
-              style={[styles.typeButton, type === 'income' && styles.activeTypeButton]}
-              onPress={() => setType('income')}
-            >
-              <Text style={[styles.typeButtonText, type === 'income' && styles.activeTypeText]}>
-                Income
-              </Text>
-            </TouchableOpacity>
-          )}
-        </View>
-
-        <View style={styles.inputContainer}>
-          <Text style={styles.label}>Amount</Text>
-          <View style={styles.amountInputContainer}>
-            <Text style={styles.currencySymbol}>฿</Text>
-            <TextInput
-              style={styles.amountInput}
-              value={amount}
-              onChangeText={setAmount}
-              keyboardType="numeric"
-              placeholder="0.00"
-              placeholderTextColor="#999"
-            />
-          </View>
-        </View>
-
-        <View style={styles.inputContainer}>
-          <Text style={styles.label}>Description (Optional)</Text>
-          <TextInput
-            style={styles.input}
-            value={description}
-            onChangeText={setDescription}
-            placeholder="Enter description"
-            placeholderTextColor="#999"
-          />
-        </View>
-
-        <TouchableOpacity
-          style={styles.dateButton}
-          onPress={() => setShowDateModal(true)}
-        >
-          <Text style={styles.dateButtonText}>
-            Date: {selectedDate.toLocaleDateString()}
-          </Text>
-        </TouchableOpacity>
-
-        <View style={styles.recurringContainer}>
-          <Text>Recurring Transaction</Text>
-          <Switch
-            value={isRecurring}
-            onValueChange={setIsRecurring}
-          />
-        </View>
-
-        {isRecurring && (
-          <View style={styles.periodSelector}>
-            {['weekly', 'monthly', 'yearly'].map(period => (
-              <TouchableOpacity
-                key={period}
-                style={[
-                  styles.periodButton,
-                  recurringPeriod === period && styles.selectedPeriod
-                ]}
-                onPress={() => setRecurringPeriod(period)}
-              >
-                <Text>{period.charAt(0).toUpperCase() + period.slice(1)}</Text>
-              </TouchableOpacity>
-            ))}
-          </View>
-        )}
-
-        <TouchableOpacity
-          style={[styles.addButton, !amount && styles.addButtonDisabled]}
-          onPress={handleAddTransaction}
-          disabled={!amount || loading}
-        >
-          {loading ? (
-            <ActivityIndicator color="#fff" />
-          ) : (
-            <Text style={styles.addButtonText}>Add Transaction</Text>
-          )}
-        </TouchableOpacity>
-      </ScrollView>
-
-      {renderDatePicker()}
-    </SafeAreaView>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-  },
-  header: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    borderBottomWidth: 1,
-    borderBottomColor: '#f0f0f0',
-  },
-  backButton: {
-    padding: 8,
-  },
-  headerTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    color: '#000',
-  },
-  rightPlaceholder: {
-    width: 40,
-  },
-  content: {
-    flex: 1,
-    padding: 16,
-  },
-  typeSelector: {
-    flexDirection: 'row',
-    marginBottom: 24,
-    backgroundColor: '#f0f0f0',
-    borderRadius: 12,
-    padding: 4,
-  },
-  typeButton: {
-    flex: 1,
-    paddingVertical: 12,
-    alignItems: 'center',
-    borderRadius: 8,
-  },
-  activeTypeButton: {
-    backgroundColor: '#8A2BE2',
-  },
-  typeButtonText: {
-    fontSize: 16,
-    color: '#666',
-    fontWeight: '600',
-  },
-  activeTypeText: {
-    color: '#fff',
-  },
-  inputContainer: {
-    marginBottom: 24,
-  },
-  label: {
-    fontSize: 16,
-    color: '#666',
-    marginBottom: 8,
-  },
-  input: {
-    backgroundColor: '#f8f8f8',
-    borderRadius: 12,
-    padding: 16,
-    fontSize: 16,
-    color: '#000',
-  },
-  amountInputContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: '#f8f8f8',
-    borderRadius: 12,
-    paddingHorizontal: 16,
-    height: 50,
-  },
-  currencySymbol: {
-    fontSize: 20,
-    color: '#8A2BE2',
-    marginRight: 8,
-  },
-  amountInput: {
-    flex: 1,
-    fontSize: 20,
-    color: '#000',
-  },
-  addButton: {
-    backgroundColor: '#8A2BE2',
-    borderRadius: 25,
-    height: 50,
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginTop: 24,
-  },
-  addButtonDisabled: {
-    backgroundColor: '#D1D1D1',
-  },
-  addButtonText: {
-    color: '#fff',
-    fontSize: 16,
-    fontWeight: 'bold',
-  },
-  dateButton: {
-    backgroundColor: '#f0f0f0',
-    borderRadius: 12,
-    padding: 16,
-    marginBottom: 24,
-  },
-  dateButtonText: {
-    fontSize: 16,
-    color: '#333',
-  },
-  recurringContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: 24,
-  },
-  periodSelector: {
-    flexDirection: 'row',
-    marginBottom: 24,
-  },
-  periodButton: {
-    flex: 1,
-    padding: 16,
-    borderWidth: 1,
-    borderColor: '#8A2BE2',
-    borderRadius: 8,
-  },
-  selectedPeriod: {
-    backgroundColor: '#8A2BE2',
-  },
-  modalOverlay: {
-    flex: 1,
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  modalContent: {
-    backgroundColor: 'white',
-    borderRadius: 20,
-    padding: 20,
-    width: '80%',
-    maxHeight: '80%',
-  },
-  modalTitle: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    marginBottom: 16,
-    textAlign: 'center',
-  },
-  dateOption: {
-    padding: 16,
-    borderBottomWidth: 1,
-    borderBottomColor: '#f0f0f0',
-  },
-  selectedDate: {
-    backgroundColor: '#8A2BE2',
-    borderRadius: 8,
-  },
-  dateText: {
-    fontSize: 16,
-    color: '#333',
-  },
-  selectedDateText: {
-    color: '#fff',
-  },
-  closeButton: {
-    marginTop: 16,
-    padding: 12,
-    backgroundColor: '#f0f0f0',
-    borderRadius: 8,
-    alignItems: 'center',
-  },
-  closeButtonText: {
-    fontSize: 16,
-    color: '#333',
-    fontWeight: '500',
-  },
-});
